requirements_dev.txt
black
streamlit 
pandas 
openpyxl
azure-storage-blob
jupyter
python-dotenv
pyarrow